componetize questionCategoryDistribution
