+++ frontend/src/components/Main/Submission/submission.jsx
@@ -6,6 +6,7 @@ import { analyzeArray } from "../../../expertArrays/analyze";
 import { evaluateArray } from "../../../expertArrays/evaluate";
 import { createArray } from "../../../expertArrays/create";
 import CsvOptions from "./CsvOptions";
+import QuestionCategoryDistribution from "./QuestionCategoryDistribution";
 
 import { uploadFile, transcribeFile } from "../../../utils/assemblyAPI";
 
@@ -577,17 +578,7 @@ export default function Submission() {
     setQuestions(sentences.filter((sentence) => sentence.isQuestion));
   }
 
-  function getAmountOfLabel(label) {
-    let amount = 0;
-    if (labeledQuestions) {
-      for (let i = 0; i < labeledQuestions.length; i++) {
-        if (labeledQuestions[i].label === label) {
-          amount++;
-        }
-      }
-      return amount;
-    }
-  }
+
 
   function getMaxSpeaker() {
     let speakTimeList1 = totalSpeakers();
@@ -873,95 +864,7 @@ export default function Submission() {
     };
   }
 
-  const barChartProps = {
-    options: {
-      title: {
-        text: "Question Category Distribution",
-        align: "left",
-        style: {
-          fontSize: "30px",
-          fontWeight: "bold",
-          fontFamily: undefined,
-          color: "#263238",
-        },
-      },
-      dataLabels: {
-        enabled: true,
-        style: {
-          fontSize: "28px",
-          fontFamily: "Helvetica, Arial, sans-serif",
-          fontWeight: "bold",
-        },
-      },
-      xaxis: {
-        labels: {
-          style: {
-            fontSize: "20px",
-          },
-        },
-        categories: [
-          "Knowledge",
-          "Understand",
-          "Apply",
-          "Analyze",
-          "Evaluate",
-          "Create",
-          "Uncategorized",
-        ],
-      },
-      tooltip: {
-        enabled: false, // Disable the tooltip on mouseover
-      },
-    },
-    series: [
-      {
-        data: [
-          {
-            x: "Knowledge",
-            y: getAmountOfLabel("Knowledge"),
-            fillColor: "#0000FF",
-            strokeColor: "#000000",
-          },
-          {
-            x: "Understand",
-            y: getAmountOfLabel("Understand"),
-            fillColor: "#D42AC8",
-            strokeColor: "#C23829",
-          },
-          {
-            x: "Apply",
-            y: getAmountOfLabel("Apply"),
-            fillColor: "#009400",
-            strokeColor: "#C23829",
-          },
-          {
-            x: "Analyze",
-            y: getAmountOfLabel("Analyze"),
-            fillColor: "#FF7300",
-            strokeColor: "#C23829",
-          },
-          {
-            x: "Evaluate",
-            y: getAmountOfLabel("Evaluate"),
-            fillColor: "#FFC400",
-            strokeColor: "#000000",
-          },
-          {
-            x: "Create",
-            y: getAmountOfLabel("Create"),
-            fillColor: "#7C7670",
-            strokeColor: "#C23829",
-          },
-          {
-            x: "Uncategorized",
-            y: getAmountOfLabel("Uncategorized"),
-            fillColor: "#FF0000",
-            strokeColor: "#C23829",
-          },
-        ],
-      },
-    ],
-  };
+
 
   const pieChartProps = {
     options: {
@@ -1672,23 +1575,7 @@ export default function Submission() {
               eventKey="questionCategoryKey"
               title="Question Category Distribution"
             >
-              <div className="card-deck mb-3 text-center">
-                <table>
-                  <tbody>
-                    <tr>
-                      <td id="barChartContainer">
-                        <Chart
-                          options={barChartProps.options}
-                          series={barChartProps.series}
-                          type="bar"
-                          width="1400"
-                          height="600"
-                        />
-                      </td>
-                    </tr>
-                  </tbody>
-                </table>
-              </div>
+              <QuestionCategoryDistribution labeledQuestions={labeledQuestions} />
             </Tab>
 
             <Tab eventKey="barChartKey" title="Talking Distribution">
